pyproject.toml
[project]
name = "aoe-scientist"
version = "0.1.0"
description = "AI research idea generator and reviewer based on AI Scientist"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.7",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.1",
    "openai>=1.3.7",
    "pandas>=2.1.3",
    "backoff>=2.2.1",
    "requests>=2.31.0",
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aoe_scientist"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ["py38"] 